Make login form accessible
Add links between labels and inputs and ensure there is sufficient
contrast for the text in the button.

Since the button text couldn't be green I also added a bit of extra
border to the button on the sides and modified the hover transition and
active states to make it a bit smoother.
