+++ app/templates/login.hbs
@@ -1,27 +1,1 @@
-<div class='login-form'>
-  {{#unless noAccountExistsError}}
-    <h2>{{t 'general.login'}}</h2>
-    {{#each error.keys as |key|}}
-      <div class='error'>{{t key}}</div>
-    {{else}}
-      <div class='error'>{{error.message}}</div>
-    {{/each}}
-    <form {{action 'authenticate' on='submit'}}>
-      <div class="item">
-        <label>{{t 'general.username'}}:</label>
-        {{input value=identification elementId='user-login-username' autocorrect="off" autocapitalize="off"}}
-      </div>
-
-      <div class="item">
-        <label>{{t 'general.password'}}:</label>
-        {{input type='password' value=password elementId='user-login-password'}}
-      </div>
-
-      <div class="buttons">
-        <button type='submit' class='done text'>{{t 'general.login'}}</button>
-      </div>
-    </form>
-  {{else}}
-    <div class='error'>{{t 'general.noAccountExists' accountName=noAccountExistsAccount}}</div>
-  {{/unless}}
-</div>
+{{login-form}}
